Shift updating a Node's signature/timestamp/size information until after the Node has been built (or visited, if it's a leaf Node).  This therefore changes the semantics of the .visited() method, so that it's called for *every* Node visited, not just leaf Nodes.
This adds a Taskmaster method for updating a Node's status without
calling the .built() and .visited() methods, which is used by the
CleanTask subclass to avoid writing .sconsign info when cleaning (-c).

Interesting ripple effect in the SConf.py module:  we have to manually
override the .store_info() method on the Nodes we create, so that they
don't write out their normal .sconsign information.  The SConf subsystem
writes out its own by hand.