Pull the NodeInfo objects out of the BuildInfo object and just deal with them as co-equal attributes of a Node.
.sconsign files now have a place-holder SConsignEntry class that holds a
Node's separate binfo and ninfo objects.  This will also be more flexible
if we ever add more information in the future.

Since this changes the format (but not content) of the .sconsign file
entries, we will need to double back and figure out how to make the
transition from pre-BSR .sconsign files before this goes live.