Set TAR_OPTIONS=--wildcards when executing rpmbuild, so the extraction of the .spec file from the tarball will work with newer versions of tar.
Change the RPM tool to initialize variables using env.SetDefault(), so
the re-application of the tools to the construction environment (?)
doesn't wipe out our $RPM setting (to add TAR_OPTIONS=--wildcards).

Have the .spec files generated by the tests build/install the software
using the version of SCons under test, not whatever happens to be
installed as "scons" on the system.

Python 1.5 fixes in various RPM code.