test/KeyboardInterrupt.py no longer hangs under Cygwin.
There seems to be a bug on Cygwin where the compiler process hangs
after sending the SIGINT signal to the process group. It is probably a
bug in cygwin1.dll, or maybe in the Python 'C' code or the Python
subprocess module. We therefore do not use 'killpg' on Cygwin.

Benoit