Updates to the new SideEffect section (kudos to Greg).
While working on this, Greg discovered a bug (issue #2154) that prevents
a SideEffect() file from being used as input to another builder call; it
makes the builder target not get build when run in paralle (e.g. -j2)...
:-(  So this patch comments out that section of Greg's section.

This also contains my usual editing pass.  In this case I changed some
of the examples and added a bunch of text to try to help clarify some
things that seemed important.  I also added a closing paragraph warning
that SideEffect() really shouldn't be used as an alternative to specifying
multiple target files in a Builder call when a command builds more than
one file that you care about.