Initial prototypes of modules for architecture support (MSCommon/arch.py) and Visual C/C++ configuration distinct from (and in preference to) the Visual Studio configuration, incorporating configuration info and ideas from Jason Kenny's latest parts/tools/MSCommon work.
This has a lot of cut-and-paste right now and does NOT actually work.
In particular, there are no changes to the rest of the infrastructure
to actually call any of this new stuff.  I'm checking these in so I can
sync on other systems for continued development.

The ultimate goal is for much of the current MSCommon/vs.py module's
logic to migrate to MSCommon/vc.py, since that's what we're actually
interested in configuring.