http://scons.tigris.org/issues/show_bug.cgi?id=2345
Fixes due to running the regression tests with the '-3' option to Python2.6,
which causes the run-time to look for potential compatibility problems with
Python 3.x.  In some cases, all we can do is quiet the warning since we still
support Python versions that can't use the newer idiom.  In other cases, we
fix the problem.  This patch contains a mix of quieting and fixing, plus a
little lint.