Increase the default timeout of TestSCons.wait_for() to 20.0 seconds. (--interactive mode on our Solaris buildbot slave takes longer than 10 seconds to create the necessary marker files.)
Set the stdin attribute on our popen object to None after we close it,
so some versions of subprocess.py don't try to write it after it's closed.

If we time out waiting for a marker file to be created, then print any
stdout or stderr that's already accumulated, since that may give clues
about why the test failed.